pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "definers"
version = "0.1.0"
authors = [
    { name = "Yaron Koresh" },
]
description = "A comprehensive Python toolkit for AI, data processing, media manipulation, and system utilities."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.26.0,<3",
    "requests>=2.31.0",
    "joblib>=1.3.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "scipy>=1.11.0,<2",
    "setuptools<81",
]

[project.license]
text = "MIT"

[project.scripts]
definers = "definers.cli:main"
definers-start = "definers.cli:main"

[project.urls]
Homepage = "https://github.com/YaronKoresh/definers"
Repository = "https://github.com/YaronKoresh/definers"
Issues = "https://github.com/YaronKoresh/definers/issues"

[project.optional-dependencies]
audio = [
    "demucs>=4.0.0",
    "librosa>=0.10.0",
    "numba>=0.57.0",
    "openai-whisper>=20231117",
    "pydub>=0.25.1",
    "soundfile>=0.12.0",
    "sox>=1.4.1",
    "torchaudio>=2.1.0",
    "resampy>=0.4.2,<0.5",
    "basic-pitch @ git+https://github.com/YaronKoresh/basic-pitch.git",
    "chatterbox-tts @ git+https://github.com/YaronKoresh/chatterbox.git",
    "madmom @ git+https://github.com/CPJKU/madmom@0551aa8f48d71a367d92b5d3a347a0cf7cd97cc9",
    "stable-ts @ git+https://github.com/jianfch/stable-ts@d89c6250fd4745115b44b83a8a3f7ebfd0e2a1f1",
]
image = [
    "imageio>=2.30.0",
    "imageio-ffmpeg>=0.4.0",
    "opencv-contrib-python-headless>=4.8.0",
    "pillow-heif>=0.13.0",
    "scikit-image>=0.21.0",
    "refiners @ git+https://github.com/finegrain-ai/refiners@d288e94fa8eed1386bd28cd0d5ceb8109c3ff398",
]
video = [
    "imageio>=2.30.0",
    "imageio-ffmpeg>=0.4.0",
    "moviepy>=1.0.3",
    "opencv-contrib-python-headless>=4.8.0",
    "scikit-image>=0.21.0",
]
ml = [
    "diffusers>=0.35.0",
    "accelerate>=1.10.0",
    "datasets>=2.14.0",
    "faiss-cpu>=1.7.4",
    "huggingface-hub>=0.20.0",
    "hydra-core>=1.3.0",
    "onnx>=1.14.0",
    "safetensors>=0.4.0",
    "scikit-learn>=1.3.0",
    "sentencepiece>=0.1.99",
    "tensorflow>=2.15.0",
    "tf-keras>=2.15.0",
    "tokenizers>=0.15.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "fairseq @ git+https://github.com/YaronKoresh/fairseq.git",
    "faiss @ https://github.com/YaronKoresh/faiss/releases/download/3b14dad/faiss-1.12.0-py3-none-manylinux_2_35_x86_64.whl ; platform_system == 'Linux' and platform_machine == 'x86_64'",
]
nlp = [
    "langdetect>=1.0.9",
    "nltk>=3.8.0",
    "sacremoses>=0.0.53",
    "stopes[inference-cpu] @ git+https://github.com/YaronKoresh/stopes.git",
]
web = [
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.100.0",
    "googledrivedownloader>=1.1.0",
    "gradio>=6.9.0",
    "gradio-client>=2.3.0",
    "lxml[html_clean]>=5.2.0",
    "matplotlib>=3.7.0",
    "playwright>=1.40.0",
]
all = [
    "definers[audio,image,video,ml,nlp,web]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.90.0",
    "expecttest>=0.3.0,<0.4.0",
    "ruff",
    "poethepoet",
    "pre-commit",
    "build",
    "vulture>=2.10",
]
cuda = [
    "cuda-python>=12.0.0",
    "nvidia-ml-py>=13.590.48",
    "distributed-ucxx-cu12==0.48.0",
    "raft-dask-cu12==26.2.0",
    "rapids-dask-dependency==26.2.0",
    "rmm-cu12==26.2.0",
    "ucx-py-cu12==0.45.0",
    "ucxx-cu12==0.48.0",
    "pylibraft-cu12==26.2.0",
    "pynvjitlink-cu12==0.7.0",
    "libucx-cu12==1.19.0",
    "cudf-cu12==26.2.1",
    "cuml-cu12==26.2.0",
    "cupy-cuda12x==14.0.1",
    "dask[complete]==2026.1.2",
    "dask-cuda==26.2.0",
    "dask-cudf-cu12==26.2.1",
    "distributed==2026.1.2",
    "nvidia-cublas-cu12>=12.8.0,<12.10.0",
    "nvidia-cuda-cupti-cu12>=12.8.0,<12.10.0",
    "nvidia-cuda-nvrtc-cu12>=12.8.0,<12.10.0",
    "nvidia-cuda-runtime-cu12>=12.8.0,<12.10.0",
    "nvidia-cudnn-cu12>=9.10.0,<9.21.0",
    "nvidia-cufft-cu12>=11.3.0,<11.5.0",
    "nvidia-curand-cu12>=10.3.0,<10.4.0",
    "nvidia-cusolver-cu12>=11.7.0,<11.8.0",
    "nvidia-cusparse-cu12>=12.5.0,<12.6.0",
    "nvidia-nccl-cu12>=2.27.0,<2.30.0",
    "nvidia-nvtx-cu12>=12.8.0,<12.10.0",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/definers",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
markers = [
    "asyncio: marks async tests",
]

[tool.ruff]
target-version = "py310"
line-length = 80

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
]
ignore = [
    "E501",
    "E402",
    "E722",
    "E741",
    "F401",
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*" = [
    "E731",
]

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.poe.tasks]
check = ["clean", "compile", "dead-code", "strip-comments", "pre-commit", "test", "clean"]

[tool.poe.tasks.test]
cmd = "pytest tests/"

[tool.poe.tasks.compile]
cmd = "python -m compileall -q ."

[tool.poe.tasks.lint]
cmd = "ruff check ."

[tool.poe.tasks.pre-commit]
cmd = "pre-commit run --all-files"

[tool.poe.tasks.pre-commit.env]
SKIP = "no-commit-to-branch"

[tool.poe.tasks.clean]
cmd = "python scripts/clean_workspace.py"

[tool.poe.tasks.strip-comments]
cmd = "python scripts/strip_comments.py"

[tool.poe.tasks.build]
cmd = "python -m build"

[tool.poe.tasks.build-wheel]
cmd = "python -m build --wheel"

[tool.poe.tasks.hook]
cmd = "pre-commit install"

[tool.poe.tasks.format]
cmd = "ruff format ."

[tool.poe.tasks.dead-code]
cmd = "vulture src/ --min-confidence 80 --sort-by-size"
